pyproject.toml
[tool.poetry]
name = "robotframework-zebrunner"
version = "0.1.0"
description = "Robotframework connector for Zebrunner reporting"
authors = ["Anatoliy Platonov <platonoff.dev@gmail.com>"]
license = "Apache"
readme = "README.md"
homepage = "https://zebrunner.com/documentation/agents/robot"
repository = "https://github.com/zebrunner/python-agent-robot"
keywords = ["automation", "zebrunner", "testing"]
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]

[tool.poetry.dependencies]
python = "^3.7"
robotframework = ">=4.0"
httpx = "^0.22.0"
pydantic = ">=1.0"
python-dotenv = ">=0.10"
PyYAML = "^5.3.1"
robotframework-seleniumlibrary = ">=4.0.0"

[tool.poetry.dev-dependencies]
mypy = "^0.942"
types-PyYAML = "^6.0.5"
robotframework-pabot = "^2.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
